backend/pyproject.toml
[tool.poetry]
name = "typst"
version = "0.1.0"
description = ""
authors = ["JMURv <architect.lock@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
djangorestframework = "^3.14.0"
django-filter = "^23.2"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.2.0"
psycopg2 = "^2.9.7"
pillow = "^10.0.0"
redis = "^4.6.0"
celery = "^5.3.1"
django-environ = "^0.10.0"
gunicorn = "^21.2.0"
djangorestframework-simplejwt = "^5.2.2"
channels = "^4.0.0"
channels-redis = "^4.1.0"
daphne = "^4.0.0"
geopy = "^2.3.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
